==> 3group/preprocessing/STEP0_retraction_reasons.ipynb <==
Connect to drive, set directory

from google.colab import drive
drive.mount('/content/drive')

cd '/content/drive/MyDrive/Thesis_UU/push/3group'

Import libraries

import pandas as pd
from collections import Counter

Load Retraction Watch database

retracted_db = pd.read_excel('Data (Excel)/Retracted_Database.xlsx')

retracted_db.rename(columns={"Record ID": "Record_ID"}, inplace = True)

retracted_db.Reason.head(5)

len(retracted_db)

retracted_db.Reason = retracted_db.Reason.str.lower() #lowercase reasons for retraction

retracted_db.Reason[1] #check data

retracted_db.Reason[12]

retracted_db[retracted_db.Reason.str.contains('error|duplication of', na=False)].head(2) #check data

retracted_db.Reason = retracted_db.Reason.str.replace(';', ',', regex = True) #clean up colun 'Reason'

retracted_db.Reason = retracted_db.Reason.str.replace('+', '', regex = True) #clean up colun 'Reason'

Splitting up reasons for retraction, as most papers have more than one reason for retraction listed

retracted_db.Reason = retracted_db.Reason.map(lambda x: str(x)[:-1]) 

results = set()
retracted_db['Reason'].str.split(',').apply(results.update)
print(results)

results #all reasons for retraction

Check how often reasons appear in the database

results = Counter()
retracted_db['Reason'].str.lower().str.split(',').apply(results.update)
print(results)
#counts number of appearance for every reason

retracted_db['Reason'].apply(lambda x: pd.value_counts(x.split(','))).sum(axis = 0).sort_values(ascending = False).head(50)
#lists 30 most frequent reasons and their appearance

retracted_db.Reason.tail(10)

## Create the 2 groups: error and misconduct

error = retracted_db[retracted_db.Reason.str.contains('error', na=False)]
#subset with every paper that contains a reason which includes the term 'error'

len(error) #size of subset

error.Reason.head(5)

error['Reason'].apply(lambda x: pd.value_counts(x.split(','))).sum(axis = 0).sort_values(ascending = False).head(30)
#lists of top 30 most frequent reasons for retraction in error group

misconduct = retracted_db[retracted_db.Reason.str.contains('fake|false|falsification|hoax|manipulation|misconduct by|paper mill|plagiarism of|randomly generated content|sabotage|salami slicing', na=False)]
#subset with every paper that contains a reason selected for misconduct group

len(misconduct) #size of subset

misconduct.Reason.head()

misconduct['Reason'].apply(lambda x: pd.value_counts(x.split(','))).sum(axis = 0).sort_values(ascending = False).head(30)
#lists of top 30 most frequent reasons for retraction in misconduct group

Remove overlapping papers / papers that contain both error and misconduct reasons

overlap = pd.merge(misconduct, error, how='inner', left_on='Record_ID', right_on='Record_ID')
#create subset containing papers with overlapping reasons

len(overlap)

overlap.head(2)

overlap.Reason_x.head(3)

error = error[~error.Record_ID.isin(overlap.Record_ID)] #removal of overlap from error subset

len(error) #size of error group

misconduct = misconduct[~misconduct.Record_ID.isin(overlap.Record_ID)] #removal of overlap from misconduct subset

len(misconduct) #size of misconduct group

error['misconduct'] = 0 #create column indicating group

error.head(2)

misconduct['misconduct'] = 1 #create column indicating group

misconduct.head(2)

Save files

error.to_csv('Data (CSV)/error_db.csv')

misconduct.to_csv('Data (CSV)/misconduct_db.csv')
