==> 3group/preprocessing/STEP1_pdf_to_text_non_retracted.ipynb <==
## Scrape non-retracted pdfs to text

Connect to drive

from google.colab import drive
drive.mount('/content/drive')

import os
files = os.listdir('.')

Set directory of the folder where the downloads are placed in

cd '/content/drive/MyDrive/Thesis_UU/push/3group/Downloads (paper pdfs)/Downloads_non_retracted'

Scrape text of pdfs, insert into dataframe

!pip install pymupdf

import pandas as pd
import fitz  # this is pymupdf
df_non_retracted = pd.DataFrame(columns=['DOI', 'Content', 'Retracted']) #make empty dataframe
for study in files: #iterate through the files
  try:
    with fitz.open(study) as doc: #open the document
        text = "" #start with empty string
        for page in doc: #iterate through the pages of the file
            text += page.get_text() #add the found text to the empty string
        df_non_retracted = df_non_retracted.append({'DOI': study, 'Content': text, 'Retracted': 0}, ignore_index=True) #add DOI, text and 0 for non-retracted paper to the dataframe
  except: #if anything above fails, pass this file
    pass

df_non_retracted.head(2)

len(df_non_retracted)

Save dataframe to a csv file

df_non_retracted.to_csv('/content/drive/MyDrive/Thesis_UU/push/3group/Data (CSV)/non_retracted_papers.csv', index=False, encoding = 'utf-8-sig')
