==> 3group/preprocessing/STEP2_retracted_merge_meta.ipynb <==
## Merge meta retracted

Connect to drive, set directory

from google.colab import drive
drive.mount('/content/drive')

cd '/content/drive/MyDrive/Thesis_UU/push/3group/'

Error

import pandas as pd

df_error = pd.read_csv('Data (CSV)/retracted_error_papers.csv')
df_meta_error = pd.read_csv('Data (CSV)/meta_error.csv')

df_error['ID'] = df_error.DOI.str.replace('.pdf', '', regex = True) #remove ".pdf" from ID

e_result = pd.merge(df_error, df_meta_error, on="ID") #merge meta and content data

len(e_result)

e_result.head(2)

e_result.Journal.value_counts().head(50) #check which journals contain how many papers 

Misconduct

df_misconduct = pd.read_csv('Data (CSV)/retracted_misconduct_papers.csv')
df_meta_misconduct = pd.read_csv('Data (CSV)/meta_misconduct.csv')

df_misconduct['ID'] = df_misconduct.DOI.str.replace('.pdf', '', regex = True) #remove ".pdf" from ID

m_result = pd.merge(df_misconduct, df_meta_misconduct, on="ID") #merge meta and content data

len(m_result)

m_result.head(2)

m_result.Journal.value_counts().head(50) #check which journals contain how many papers 

### Creating subsets containing only journals that contain more than one paper in both error and misconduct group

journals_e = e_result.Journal.value_counts() #create df with journals and amount of papers from error group
e_filtered = e_result[e_result.Journal.isin(journals_e.index[journals_e.gt(1)])] #remove all journals which only contain one journal

e_filtered.Journal.value_counts().tail(2)

journals_e_set = set(e_filtered['Journal'])
print(len(journals_e_set)) #amount of journals
print(journals_e_set) #names of journals

sub_misconduct = m_result[m_result.Journal.isin(journals_e_set)] #create subset of misconduct data only with journals from error data

len(sub_misconduct)

sub_misconduct.head(2)

journals_m = sub_misconduct.Journal.value_counts() #create df with journals and amount of papers from misconduct group
sub_misconduct = sub_misconduct[sub_misconduct.Journal.isin(journals_m.index[journals_m.gt(1)])] #remove all journals which only contain one journal

sub_misconduct.Journal.value_counts().tail(2)

journals_m_set = set(sub_misconduct['Journal'])
print(len(journals_m_set)) #amount of journals
print(journals_m_set) #name of journals

sub_error = e_result[e_result.Journal.isin(journals_m_set)] #subsetting error data to only contain journals which appear in misconduct data

len(sub_error)

sub_error.head(2)

sub_retracted = pd.concat([sub_error, sub_misconduct]) #concat to one dataset containing both error and misconduct data

len(sub_retracted) #length of created subset containing both groups

len(set(sub_retracted['Journal'])) #amount of journals

retracted_journals = sub_retracted.groupby(['Journal']).size().reset_index(name='Amount').sort_values(by = 'Amount', ascending = False)
retracted_journals #check which journals contain how many papers 

retracted_journals.to_csv('Data (CSV)/retracted_journals.csv') #save journal df to csv

sub_retracted.to_csv('Data (CSV)/sub_retracted.csv') #save merged data to csv
