==> 3group/preprocessing/STEP4_non_retracted_merge_meta.ipynb <==
## Merge meta non-retracted

Connect to drive, set directory

from google.colab import drive
drive.mount('/content/drive')

cd '/content/drive/MyDrive/Thesis_UU/push/3group/'

Non-retracted

import pandas as pd

df_non_retracted = pd.read_csv('Data (CSV)/non_retracted_papers.csv') 
df_meta_non_retracted = pd.read_csv('Data (CSV)/meta_non_retracted.csv')

df_non_retracted['ID'] = df_non_retracted.DOI.str.replace('.pdf', '', regex = True) #remove ".pdf" from ID

nr_result = pd.merge(df_non_retracted, df_meta_non_retracted, on="ID") #merge meta and content data

len(nr_result)

nr_result.head(2)

nr_result.tail(2)

nr_result.columns

nr_result.rename(columns={"Source Title": "Source_Title"}, inplace = True)

len(set(nr_result['Source_Title'])) #amount of journals

non_retracted_journals = nr_result.groupby(['Source_Title']).size().reset_index(name='Amount').sort_values(by = 'Amount', ascending = False)
non_retracted_journals #check which journals contain how many papers 

non_retracted_journals.to_csv('Data (CSV)/non_retracted_journals.csv') #save journal df to csv

nr_result.to_csv('Data (CSV)/non_retracted.csv') #save merged data to csv
