==> 3group/preprocessing/STEP7_journal_distributions.ipynb <==
Connect to drive, set directory

from google.colab import drive
drive.mount('/content/drive')

cd '/content/drive/MyDrive/Thesis_UU/push/3group'

Import Libraries

import pandas as pd
import re

## External validation dataset

journal_test_data_set = pd.read_csv('Data (CSV)/journal_test_data_set.csv')

len(journal_test_data_set)

Investigate group sizes

classes = journal_test_data_set.groupby(['class']).count().reset_index()
classes

len(set(journal_test_data_set.Journal_Name)) #amount of journals

Create df of journals with group size per journals

journals_test = journals_test.unstack('class').reset_index()

journals_test.rename(columns={"Journal_Name": "journal", "e": "error", "m": "misconduct", "nr": "non-retracted"}, inplace = True)

journals_test['total']= journals_test.sum(axis=1)

journals_test

Save df

journals_test.to_csv('Data (CSV)/journals_test.csv')

## Train/test dataset

journal_train_data_set = pd.read_csv('Data (CSV)/journal_train_data_set.csv')

len(journal_train_data_set)

Investigate group sizes

classes = journal_train_data_set.groupby(['class']).count().reset_index()
classes

len(set(journal_train_data_set.Journal_Name)) #amount of journals

Create df of journals with group size per journals

journals_train = journals_train.unstack('class').reset_index()

journals_train.rename(columns={"Journal_Name": "journal", "e": "error", "m": "misconduct", "nr": "non-retracted"}, inplace = True)

journals_train['total']= journals_train.sum(axis=1)

journals_train

Save df

journals_train.to_csv('Data (CSV)/journals_train.csv')
