==> 3group/preprocessing/STEP8_sampling_by_journal.ipynb <==
## Balancing classes stratisified by journal

Connect to drive, set directory

from google.colab import drive
drive.mount('/content/drive')

cd '/content/drive/MyDrive/Thesis_UU/push/3group'

Import libraries

import pandas as pd

Load train/test dataset

train_dataset = pd.read_csv('Data (CSV)/journal_train_data_set.csv', encoding="utf-8-sig")

train_dataset.head(2)

len(train_dataset)

train_dataset.drop(columns=['Unnamed: 0', 'Unnamed: 0.1'], inplace = True)

Df with number of papers of largest group (non-retracted) per journal

journals_nr = train_dataset[train_dataset['class'] == 'nr'].groupby(['Journal_Name']).size().reset_index(name='amount')
journals_nr

len(train_dataset[train_dataset['class'] == 'nr'])

train_dataset.columns

i = 1
journals_nr.iloc[i, 1] #number

journals_nr.iloc[i, 0] #journal name

#create empty dataframe
sample_papers = pd.DataFrame(columns = ['ID', 'All content', 'Text', 'Journal_Name', 'class', 'Retracted', 'preprocessed'])

#create empty list
samples = []

journals_e.shape[0] #amount of journals in total

Sample papers from other groups to match largest, per journal: if a journal contains less samples from other groups than non-retracted, oversamples other groups, if not, samples other group

for i in range(0, len(journals_nr)):
  j = journals_nr.iloc[i, 0] #journal name
  a = journals_nr.iloc[i, 1] #number of papers

  if len(train_dataset[(train_dataset['class'] == 'e') & (train_dataset['Journal_Name'] == j)]) < a:
    sample_e = train_dataset[(train_dataset['class'] == 'e') & (train_dataset['Journal_Name'] == j)].sample(a, random_state = 1, replace=True)
  else:
    sample_e = train_dataset[(train_dataset['class'] == 'e') & (train_dataset['Journal_Name'] == j)].sample(a, random_state = 1)
  
  if len(train_dataset[(train_dataset['class'] == 'm') & (train_dataset['Journal_Name'] == j)]) < a:
    sample_m = train_dataset[(train_dataset['class'] == 'm') & (train_dataset['Journal_Name'] == j)].sample(a, random_state = 1, replace=True)
  else:
    sample_m = train_dataset[(train_dataset['class'] == 'm') & (train_dataset['Journal_Name'] == j)].sample(a, random_state = 1)

  sample_nr = train_dataset[(train_dataset['class'] == 'nr') & (train_dataset['Journal_Name'] == j)]
  #appends papers per journal for every group
  samples.append(sample_e)
  samples.append(sample_m)
  samples.append(sample_nr)

len(samples) #nested list

for i in samples:
  sample_papers = pd.concat([i, sample_papers]) #unnests list

len(sample_papers) #amount of total papers, after sampling

sample_papers.head()

Saving files

sample_papers.to_csv('Data (CSV)/papers_journal_balanced.csv')
